Disable roxygen comments processing for unavailable methods
I had been looking at outdated sample code which included a use of "create_dot_body" which has been used as a sub-method for plot.grf_tree since. As "create_dot_body" was listed in the 00index.html, I was under the assumption it should be available as a method even though it was not. Maybe this will help prevent confusion of that kind while still maintaining the commentary. Would this comply with the GRF coding/documentation practices?