Consolidate the observations and data objects. (#363)
Previously, a special object was created to hold the observation information, and this
object would be serialized along with the forest. Now observations are simply exposed
as part of the data object, through the methods Data#get_outcome, get_treatment, and
get_instrument. This refactor has a few advantages:

* It results in a simpler and more consistent code structure.
* LocalLinearPredictionStrategy no longer needs to store the data matrices.
* It cuts down on unnecessary serialization/ deserialization.

Note that with this change, the original training data must now be passed to prediction.
The R wrapper methods predict.causal_forest, predict.regression_forest were updated
to reconstruct the train matrix before calling into the core prediction methods.