Make sure R tests are included in package builds. (#432)
This PR updates the package build script script to ensure that all tests are
run. Now CI will run all R tests in the package, as opposed to a light smoke
test as was run previously.

With this PR, `build_package.R` will include and run all tests in the package
by default. A new option `--as-cran` builds the package in preparation for CRAN
submission, excluding all tests except a simple smoke test.

The PR also fixes several tests that fail on linux to ensure the package passes
CI.