Add lintr to CI and bring existing R source up to standard (#437)
This PR adds lintr as the first check in build_package.R.

The style is the default "tidyverse" R standard with the exception of 5 options disabled for potential false positives:

* line_length_linter(120)
* object_name_linter = NULL
* commented_code_linter = NULL
* object_usage_linter = NULL
* spaces_left_parentheses_linter = NULL

RStudio can make a file conform to this in: Addins->Styler.

This PR also brings the existing R source up to this standard by touching a large number of files under r-package (done automatically with styler::style_pkg()). See PR #437 for details.