Add option to disable pruning for honest trees (#456)
This PR extends the honesty parameters with a boolean flag
prune_empty_leaves / prune.empty.leaves that toggles whether the honest
tree is pruned or not (this was hardcoded to always be true). It is
added as an experimental option to tune performance of honest
trees on small data sets.