Run the characterization tests in CI.
Previously, the C++ characterization tests only passed when using clang because
of differences in the way random numbers are generated across platforms.
Because we build on a couple different platforms, we had to disable these tests
in CI.

Now that we've added platform-independent random number generation in #469, we
can enable the characterization tests.