Remove most references to shared_ptr. (#508)
This commit replaces most uses of 'shared_ptr' with 'unique_ptr', invoking
'move' when necessary to transfer ownership. In general, it is best to use
'unique_ptr' since it encourages a clear design around object ownership.

We still use shared_ptr to hold the list of trees -- this will be addressed in
a follow-up.