Revert "Make collect_predictions_batch return a unique_ptr to vector<Prediction>"
This reverts commit 3d0c763eec073c4eda01864d3752c410ece9b4b6.

Returning a pointer to vector<Prediction> is not necessary in C++11
because of move semantics (compiler will elide copy).
(https://stackoverflow.com/questions/15704565/efficient-way-to-return-a-stdvector-in-c)

The question of whether this is the best aggregation of all the
predictions remains:

There is one vector<Prediction> of size num_samples allocated, and there are
num_threads * num_samples_in_thread vector<Prediction> objects
allocated, i.e. twice the amount of the serial version?