Treat warnings as errors in testthat (#595)
One might introduce a coding error somewhere such that an existing test
that passed without warning, now gives a warning (such as for example
most average treatment effect functions, and tuning functions).

This PR enables warnings as errors to avoid this. It also updates tests
to remove the 4 current warnings.