Update data.get_all_values to return sort indices
Everything is the same except that a vector<size_t> of node sample IDs
corresponding to the sort order is returned

Technically one could omit the extra "sorted_samples" vector and just
return the index vector and index with `samples[index[i]]`.

This is more readable and can be updated if performance sensitive

Note that: sorted_samples is std::moved back.
values are copied into all_values because this makes sorting faster
because of cache coherency.