Assign random number seed independent of num_threads
`ForestTrainer::train_trees` calls `split_sequence` to split the number of trees to grow into batches depending on the specified number of threads. However, when assigning a seed to grow an individual tree, only the batch index is considered, but not the tree index. Therefore, different values for the number of threads will yield different results, despite the user having specified the same (global) random number seed.

The patch initializes the random number generator based on the user provided seed and the index of the tree (`start + i`). Thus, each tree will be assigned the same seed, independent of the specified number of threads.